% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-glycan-descriptions.R
\name{add_comp_descriptions}
\alias{add_comp_descriptions}
\title{Add Descriptions to Glycan Compositions}
\usage{
add_comp_descriptions(exp)
}
\arguments{
\item{exp}{An \code{\link[=experiment]{experiment()}} object.}
}
\value{
The experiment object with the new columns added.
}
\description{
This function adds the following columns to the
variable information tibble:
\itemize{
\item 'n_hex': number of Hex
\item 'n_hexnac': number of HexNAc
\item 'n_fuc': number of Fuc
\item 'n_neuac': number of NeuAc
\item 'n_neugc': number of NeuGc
\item 'n_sia': number of Sia
}
}
\details{
\code{n_sia} is the sum of \code{n_neuac} and \code{n_neugc}.
}
