% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-glycan-descriptions.R
\name{add_struct_descriptions}
\alias{add_struct_descriptions}
\title{Add Descriptions to Glycan Structures}
\usage{
add_struct_descriptions(exp)
}
\arguments{
\item{exp}{An \code{\link[=experiment]{experiment()}} object.}
}
\value{
The experiment object with the new columns added.
}
\description{
This function adds columns about glycan structural properties
to the variable information tibble.
Depending on the glycan type (N-glycan, O-glycan),
different columns are added.
Current, only N-glycan descriptions are implemented.
See \code{\link[glymotif:describe_n_glycans]{glymotif::describe_n_glycans()}} for columns added.
}
\details{
To use this function, the \code{\link[=experiment]{experiment()}} object must contain
a \code{glycan_structure} column in the \code{var_info} tibble.
If \code{add_structures} has already been called,
the cached structures are used.
If not, \code{add_structures} is called internally,
and parsed structures are added to the \code{\link[=experiment]{experiment()}} object.
}
\seealso{
\code{\link[=add_structures]{add_structures()}}, \code{\link[glymotif:describe_n_glycans]{glymotif::describe_n_glycans()}}
}
