% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-pseudo-glycome.R
\name{as_pseudo_glycomes}
\alias{as_pseudo_glycomes}
\title{Convert a glycoproteomics experiment to pseudo-glycomes}
\usage{
as_pseudo_glycomes(exp, aggr_method = c("sum", "mean", "median"))
}
\arguments{
\item{exp}{A \code{\link[=GlycoproteomicSE]{GlycoproteomicSE()}}.}

\item{aggr_method}{Aggregation method to use. One of "sum", "mean", or
"median". Default is "sum".}
}
\value{
A named list of \code{\link[=GlycomicSE]{GlycomicSE()}} objects. List names identify each
glycosite as \code{{protein}-{protein_site}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Converts a \code{\link[=GlycoproteomicSE]{GlycoproteomicSE()}} into one \code{\link[=GlycomicSE]{GlycomicSE()}} for each complete
\verb{(protein, protein_site)} pair in its row data.
}
