% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_compositions}
\alias{count_compositions}
\alias{count_structures}
\alias{count_peptides}
\alias{count_glycopeptides}
\alias{count_glycoforms}
\alias{count_proteins}
\alias{count_glycosites}
\title{Identification overview}
\usage{
count_compositions(x)

count_structures(x)

count_peptides(x)

count_glycopeptides(x, count_struct = NULL)

count_glycoforms(x, count_struct = NULL)

count_proteins(x)

count_glycosites(x)
}
\arguments{
\item{x}{An \code{\link[=experiment]{experiment()}} object.}

\item{count_struct}{For \code{count_glycopeptides()} and \code{count_glycoforms()},
whether to count the number of glycan structures or glycopeptides.
If \code{TRUE}, glycopeptides or glycoforms bearing different glycan structures
with the same glycan composition are counted as different ones.
If not provided (NULL), defaults to \code{TRUE} if \code{glycan_structure} column exists
in the variable information tibble, otherwise \code{FALSE}.}
}
\value{
An integer.
\itemize{
\item \code{count_compositions()}: The number of glycan compositions.
\item \code{count_structures()}: The number of glycan structures.
\item \code{count_peptides()}: The number of peptides.
\item \code{count_glycopeptides()}: The number of unique combinations of peptides, sites, and glycans.
\item \code{count_glycoforms()}: The number of unique combinations of proteins, sites, and glycans.
\item \code{count_proteins()}: The number of proteins.
\item \code{count_glycosites()}: The number of unique combinations of proteins and sites.
}
}
\description{
These functions are used to identify the number of
glycan compositions, glycan structures, glycopeptides, peptides,
glycoforms, glycoproteins, and glycosites in an \code{\link[=experiment]{experiment()}}.
}
\details{
The following columns are required in the variable information tibble:
\itemize{
\item \code{count_compositions()}: \code{glycan_composition}
\item \code{count_structures()}: \code{glycan_structure}
\item \code{count_peptides()}: \code{peptide}
\item \code{count_glycopeptides()}: \code{glycan_composition} or \code{glycan_structure}, \code{peptide}, \code{peptide_site}
\item \code{count_glycoforms()}: \code{glycan_composition} or \code{glycan_structure},
\code{protein} or \code{proteins}, \code{protein_site} or \code{protein_sites}
\item \code{count_proteins()}: \code{protein} or \code{proteins}
\item \code{count_glycosites()}: \code{protein} or \code{proteins}, \code{protein_site} or \code{protein_sites}
}

You can use \code{count_struct} parameter to control how to count glycopeptides and glycoforms,
either by glycan structures or by glycan compositions.
}
\examples{
exp <- real_experiment
count_compositions(exp)
count_structures(exp)
count_peptides(exp)
count_glycopeptides(exp)
count_glycoforms(exp)
count_proteins(exp)
count_glycosites(exp)

}
