% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-zzz-deprecated.R
\name{deprecated-dplyr-aliases}
\alias{deprecated-dplyr-aliases}
\alias{arrange_obs}
\alias{arrange_var}
\alias{filter_obs}
\alias{filter_var}
\alias{mutate_obs}
\alias{mutate_var}
\alias{rename_obs}
\alias{rename_var}
\alias{select_obs}
\alias{select_var}
\alias{slice_obs}
\alias{slice_var}
\alias{slice_head_obs}
\alias{slice_head_var}
\alias{slice_tail_obs}
\alias{slice_tail_var}
\alias{slice_sample_obs}
\alias{slice_sample_var}
\alias{slice_max_obs}
\alias{slice_max_var}
\alias{slice_min_obs}
\alias{slice_min_var}
\alias{left_join_obs}
\alias{left_join_var}
\alias{inner_join_obs}
\alias{inner_join_var}
\alias{semi_join_obs}
\alias{semi_join_var}
\alias{anti_join_obs}
\alias{anti_join_var}
\title{Deprecated dplyr-style aliases}
\usage{
arrange_obs(...)

arrange_var(...)

filter_obs(...)

filter_var(...)

mutate_obs(...)

mutate_var(...)

rename_obs(...)

rename_var(...)

select_obs(...)

select_var(...)

slice_obs(...)

slice_var(...)

slice_head_obs(...)

slice_head_var(...)

slice_tail_obs(...)

slice_tail_var(...)

slice_sample_obs(...)

slice_sample_var(...)

slice_max_obs(...)

slice_max_var(...)

slice_min_obs(...)

slice_min_var(...)

left_join_obs(...)

left_join_var(...)

inner_join_obs(...)

inner_join_var(...)

semi_join_obs(...)

semi_join_var(...)

anti_join_obs(...)

anti_join_var(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

The \verb{_obs()} and \verb{_var()} helpers have been renamed to \verb{_col()} and \verb{_row()},
respectively. Use the new names in new code.
}
\keyword{internal}
