% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rownames-colnames.R
\name{dimnames.glyexp_experiment}
\alias{dimnames.glyexp_experiment}
\title{Dimname for Experiment}
\usage{
\method{dimnames}{glyexp_experiment}(x, ...)
}
\arguments{
\item{x}{An \code{\link[=experiment]{experiment()}}.}

\item{...}{Ignored.}
}
\value{
A list with the dimnames of the expression matrix.
}
\description{
The dimnames method for \code{\link[=experiment]{experiment()}} objects are
the dimnames of their expression matrix.
}
