% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-site-aa-pos-non-na.R
\name{.compute_site_aa_pos_non_na}
\alias{.compute_site_aa_pos_non_na}
\title{Compute \if{html}{\out{<aa>}}\if{html}{\out{<pos>}} for non-NA protein_site values}
\usage{
.compute_site_aa_pos_non_na(var_info, fasta = NULL, taxid = 9606)
}
\description{
Compute \if{html}{\out{<aa>}}\if{html}{\out{<pos>}} for non-NA protein_site values
}
\keyword{internal}
