% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize-variable.R
\name{.ensure_unique}
\alias{.ensure_unique}
\title{Ensure variable IDs are unique by adding numeric suffixes}
\usage{
.ensure_unique(vars, unique_suffix)
}
\description{
Ensure variable IDs are unique by adding numeric suffixes
}
\keyword{internal}
