% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize-variable.R
\name{.fetch_uniprot_sequences}
\alias{.fetch_uniprot_sequences}
\title{Fetch protein sequences from UniProt in batches}
\usage{
.fetch_uniprot_sequences(proteins, taxid = 9606, batch_size = 50)
}
\description{
Fetch protein sequences from UniProt in batches
}
\keyword{internal}
