% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize-variable.R
\name{.get_aa_from_fasta}
\alias{.get_aa_from_fasta}
\title{Get amino acid from FASTA sequences at protein_site position}
\usage{
.get_aa_from_fasta(var_info, fasta)
}
\description{
Get amino acid from FASTA sequences at protein_site position
}
\keyword{internal}
