% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-aa-from-peptide.R
\name{.get_aa_from_peptide}
\alias{.get_aa_from_peptide}
\title{Get amino acid from peptide sequence at peptide_site position}
\usage{
.get_aa_from_peptide(var_info)
}
\description{
Get amino acid from peptide sequence at peptide_site position
}
\keyword{internal}
