% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-aa-from-uniprot.R
\name{.get_aa_from_uniprot}
\alias{.get_aa_from_uniprot}
\title{Get amino acid from UniProt sequences}
\usage{
.get_aa_from_uniprot(var_info, taxid = 9606)
}
\description{
Get amino acid from UniProt sequences
}
\keyword{internal}
