% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize-variable.R
\name{.get_default_format}
\alias{.get_default_format}
\title{Get default format string based on exp_type}
\usage{
.get_default_format(exp_type, var_info)
}
\description{
Get default format string based on exp_type
}
\keyword{internal}
