% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize-variable.R
\name{.glue_with_composition}
\alias{.glue_with_composition}
\title{Format variable strings with special handling for glycan_composition}
\usage{
.glue_with_composition(var_info, format)
}
\description{
Format variable strings with special handling for glycan_composition
}
\keyword{internal}
