% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize-variable.R
\name{.resolve_site_token}
\alias{.resolve_site_token}
\title{Replace \if{html}{\out{<site>}} token in format string with computed site values}
\usage{
.resolve_site_token(var_info, format, site_aa_pos)
}
\description{
Replace \if{html}{\out{<site>}} token in format string with computed site values
}
\keyword{internal}
