% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{filter_samples}
\alias{filter_samples}
\title{Filter samples based on conditions}
\usage{
filter_samples(exp, ...)
}
\arguments{
\item{exp}{The \link{Experiment} object.}

\item{...}{Conditions for filtering samples, passed to \code{\link[dplyr:filter]{dplyr::filter()}}.}
}
\value{
The \link{Experiment} object.
}
\description{
Filter samples based on conditions specified in the \code{\link[dplyr:filter]{dplyr::filter()}} function.
For example, \code{filter_samples(exp, group == "A")} will keep samples in \code{exp}
with "group" equal to "A".
}
\details{
Instead of \code{Experiment$filter_samples()}, which updates the \link{Experiment} object in place,
this function returns a new \code{Experiment} object with samples filtered.
The original \code{Experiment} object will not be altered.
Besides, as the first argument is the \code{Experiment} object,
this function could be used in a \code{magrittr} pipe.
}
