% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_variables}
\alias{filter_variables}
\title{Filter variables based on conditions}
\usage{
filter_variables(exp, ...)
}
\arguments{
\item{exp}{The \link{Experiment} object.}

\item{...}{Conditions for filtering variables, passed to \code{\link[dplyr:filter]{dplyr::filter()}}.}
}
\value{
The \link{Experiment} object.
}
\description{
Filter variables based on conditions specified in the \code{\link[dplyr:filter]{dplyr::filter()}} function.
For example, \code{filter_variables(exp, type == "B")} will keep variables in \code{exp}
with "type" equal to "B".
}
\details{
Instead of \code{Experiment$filter_variables()}, which updates the \link{Experiment} object in place,
this function returns a new \code{Experiment} object with variables filtered.
The original \code{Experiment} object will not be altered.
Besides, as the first argument is the \code{Experiment} object,
this function could be used in a \code{magrittr} pipe.
}
