% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-getters.R
\name{get_glycan_structures}
\alias{get_glycan_structures}
\title{Get glycan structures from an \code{\link[=experiment]{experiment()}}}
\usage{
get_glycan_structures(exp)
}
\arguments{
\item{exp}{An \code{\link[=experiment]{experiment()}}.}
}
\value{
A named list of glycan structures.
}
\description{
You need to first call \code{\link[=add_structures]{add_structures()}} to parse the glycan structures
and store them in the \code{glycan_structures} attribute of the experiment object.
}
