% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-data.R
\name{get_meta_data}
\alias{get_meta_data}
\alias{get_exp_type}
\alias{get_glycan_type}
\title{Get the meta data of an experiment}
\usage{
get_meta_data(exp, x = NULL)

get_exp_type(exp)

get_glycan_type(exp)
}
\arguments{
\item{exp}{An \code{\link[=experiment]{experiment()}}.}

\item{x}{A string, the name of the meta data field.
If \code{NULL} (default), a list of all meta data fields will be returned.}
}
\value{
The value of the meta data field. If the field does not exist,
\code{NULL} will be returned.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These metadata helpers were deprecated with \code{experiment()}. Use
\code{S4Vectors::metadata()} on a \code{SummarizedExperiment} instead.

Meta data is some descriptions about the experiment,
like the experiment type ("glycomics" or "glycoproteomics"),
or the glycan type ("N", "O", "O-GalNAc", "O-Man", "O-Fuc",
"O-GlcNAc", "O-Glc", "HMO", "GSL", "GAG", or "GPI").
}
\examples{
get_meta_data(real_experiment)
get_exp_type(real_experiment)
get_glycan_type(real_experiment)
}
\keyword{internal}
