% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-getters.R
\name{get_var_info}
\alias{get_var_info}
\title{Get the variable information of an experiment}
\usage{
get_var_info(exp)
}
\arguments{
\item{exp}{An \code{\link[=experiment]{experiment()}}.}
}
\value{
A tibble of variable information.
}
\description{
A \code{tibble} of variable information, with the first column being "variable".

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This legacy API is retained temporarily for compatibility. Use
\code{\link[=GlycomicSE]{GlycomicSE()}} or \code{\link[=GlycoproteomicSE]{GlycoproteomicSE()}} as the default data container.
}
\keyword{internal}
