% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyexp-package.R
\docType{package}
\name{glyexp-package}
\alias{glyexp}
\alias{glyexp-package}
\title{glyexp: Glycoproteomics and Glycomics Experiments}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides 'SummarizedExperiment'-based containers for glycomics and glycoproteomics experimental data. 'GlycomicSE' and 'GlycoproteomicSE' validate glycan-specific annotations while retaining standard assay, 'rowData', 'colData', and metadata interfaces. The legacy 'experiment()' API remains available with lifecycle deprecation warnings to support a staged migration of the 'glycoverse' ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://glycoverse.github.io/glyexp/}
  \item \url{https://github.com/glycoverse/glyexp}
  \item Report bugs at \url{https://github.com/glycoverse/glyexp/issues}
}

}
\author{
\strong{Maintainer}: Bin Fu \email{23110220018@m.fudan.edu.cn} (\href{https://orcid.org/0000-0001-8567-2997}{ORCID}) [copyright holder]

}
\keyword{internal}
