% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-experiment.R
\name{is_experiment}
\alias{is_experiment}
\title{Check if an Object is an Experiment}
\usage{
is_experiment(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
A logical value.
}
\description{
This function checks if an object is an experiment,
i.e. if it inherits from the class \code{glyexp_experiment}.
}
