% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate_samples}
\alias{mutate_samples}
\title{Mutate samples based on conditions}
\usage{
mutate_samples(exp, ...)
}
\arguments{
\item{exp}{The \link{Experiment} object.}

\item{...}{Name-value pairs, passed to \code{\link[dplyr:mutate]{dplyr::mutate()}}.}
}
\value{
The new \link{Experiment} object.
}
\description{
Mutate the sample information tibble using the \code{\link[dplyr:mutate]{dplyr::mutate()}} function.
For example, \verb{mutate_samples(new_group = if_else(new_col = 1)}
will add a new column "new_group" to the sample information tibble.
}
\details{
Instead of \code{Experiment$mutate_samples()}, which updates the \link{Experiment} object in place,
this function returns a new \code{Experiment} object with sample_info tibble updated.
The original \code{Experiment} object will not be altered.
Besides, as the first argument is the \code{Experiment} object,
this function could be used in a \code{magrittr} pipe.
}
