% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate_variables}
\alias{mutate_variables}
\title{Mutate variables based on conditions}
\usage{
mutate_variables(exp, ...)
}
\arguments{
\item{exp}{The \link{Experiment} object.}

\item{...}{Name-value pairs, passed to \code{\link[dplyr:mutate]{dplyr::mutate()}}.}
}
\value{
The new \link{Experiment} object.
}
\description{
Mutate the variable information tibble using the \code{\link[dplyr:mutate]{dplyr::mutate()}} function.
For example, \verb{mutate_variables(exp, new_group = if_else(new_col = 1)}
will add a new column "new_group" to the variable information tibble of \code{exp}.
}
\details{
Instead of \code{Experiment$mutate_variables()}, which updates the \link{Experiment} object in place,
this function returns a new \code{Experiment} object with variable_info tibble updated.
The original \code{Experiment} object will not be altered.
Besides, as the first argument is the \code{Experiment} object,
this function could be used in a \code{magrittr} pipe.

The columns "variable" is protected from being modified or renamed.
It is essential for keeping the linkage between the variable information and the expression matrix.
You may add new columns by transforming the "variable" column.
}
