% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-and-variables.R
\name{n_samples}
\alias{n_samples}
\alias{n_variables}
\title{Get number of samples or variables of an experiment}
\usage{
n_samples(exp)

n_variables(exp)
}
\arguments{
\item{exp}{An \code{\link[=experiment]{experiment()}}.}
}
\value{
An integer with the number of samples or variables.
}
\description{
Getting the number of samples or variables of an \code{\link[=experiment]{experiment()}}.
Syntax sugar for \code{ncol(exp$expr_mat)} and \code{nrow(exp$expr_mat)}.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This legacy API is retained temporarily for compatibility. Use
\code{\link[=GlycomicSE]{GlycomicSE()}} or \code{\link[=GlycoproteomicSE]{GlycoproteomicSE()}} as the default data container.
}
\examples{
exp <- toy_experiment
n_samples(exp)
n_variables(exp)

}
\keyword{internal}
