% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-experiment.R
\name{print.glyexp_experiment}
\alias{print.glyexp_experiment}
\title{Print a legacy experiment}
\usage{
\method{print}{glyexp_experiment}(x, ...)
}
\arguments{
\item{x}{A legacy experiment.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This method was deprecated with \code{experiment()}. Use the \code{show()} method for
a \code{\link[=GlycomicSE]{GlycomicSE()}} or \code{\link[=GlycoproteomicSE]{GlycoproteomicSE()}} object instead.
}
\keyword{internal}
