% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{real_experiment}
\alias{real_experiment}
\title{Real glycoproteomics experiment}
\format{
An \code{\link[=experiment]{experiment()}} object with 4262 variables and 12 samples.
\subsection{Variable information:}{
\itemize{
\item \code{peptide}: peptide sequence
\item \code{peptide_site}: site position on the peptide
\item \code{protein}: protein accession
\item \code{protein_site}: site position on the protein
\item \code{gene}: gene symbol
\item \code{glycan_composition}: glycan composition (glyrepr::glycan_composition())
\item \code{glycan_structure}: glycan structure (glyrepr::glycan_structure())
}
}

\subsection{Sample information:}{
\itemize{
\item \code{sample}: sample ID
\item \code{group}: disease group, one of "H" (healthy), "M" (hepatitis), "Y" (cirrhosis), and "C" (hepatocellular carcinoma)
}
}
}
\usage{
real_experiment
}
\description{
A real glycoproteomics experiment object.
This dataset is derived from the unpublished ProteomeXchange dataset PXD063749.
It contains serum N-glycoproteome profiles from 12 patients with different liver conditions:
healthy (H), hepatitis (M), cirrhosis (Y), and hepatocellular carcinoma (C).
Glycopeptide identification and annotation were performed using pGlyco3
(https://github.com/pFindStudio/pGlyco3), and quantification was carried out with pGlycoQuant
(https://github.com/Power-Quant/pGlycoQuant).
The raw data were imported with \code{glyread::read_pglyco3_pglycoquant()}.
}
\keyword{datasets}
