% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{real_experiment2}
\alias{real_experiment2}
\title{Real Glycomics Experiment}
\format{
An \code{\link[=experiment]{experiment()}} object with 67 variables and 144 samples.
\subsection{Variable information:}{
\itemize{
\item \code{glycan_composition}: glycan composition (glyrepr::glycan_composition())
\item \code{glycan_structure}: glycan structure (glyrepr::glycan_structure())
}
}

\subsection{Sample information:}{
\itemize{
\item \code{sample}: sample ID
\item \code{group}: disease group, one of "H" (healthy), "M" (hepatitis), "Y" (cirrhosis), and "C" (hepatocellular carcinoma)
}
}
}
\usage{
real_experiment2
}
\description{
A real glycomics experiment object.
This dataset is derived from the unpublished glycoPOST dataset GPST000313.
It contains serum N-glycome profiles from 144 samples with different liver conditions:
healthy (H), hepatitis (M), cirrhosis (Y), and hepatocellular carcinoma (C).
}
\keyword{datasets}
