% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\title{Rename an Experiment}
\usage{
rename(exp, new_name)
}
\arguments{
\item{exp}{An experiment object.}

\item{new_name}{A character string with the new name.}
}
\value{
An experiment object with the new name.
}
\description{
Getting a new experiment with a new name.
}
