% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-rename.R
\name{rename_samples}
\alias{rename_samples}
\alias{rename_variables}
\title{Rename Columns in the Sample or Variable Information Tibble}
\usage{
rename_samples(exp, ...)

rename_variables(exp, ...)
}
\arguments{
\item{exp}{An \code{\link[=experiment]{experiment()}}.}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Name pairs to rename.
Use \code{new_name = old_name} to rename columns.}
}
\value{
An new \code{\link[=experiment]{experiment()}} object.
}
\description{
These two functions provide a way to rename columns in the sample or variable
information tibble of an \code{\link[=experiment]{experiment()}}.

The same syntax as \code{dplyr::rename()} is used.
For example, to rename the "group" column in the sample information tibble to "condition",
use \code{rename_samples(exp, condition = group)}.
Note that you can't rename the "sample" column in the sample information tibble,
as well as the "variable" column in the variable information tibble.
These two columns are used to link the sample or variable information tibble
to the expression matrix.
}
\examples{
toy_exp <- toy_experiment()
toy_exp

# Rename columns in sample information tibble
rename_samples(toy_exp, condition = group)

# Rename columns in variable information tibble
rename_variables(toy_exp, composition = glycan_composition)

}
