% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_samples}
\alias{select_samples}
\title{Select columns of the sample information tibble}
\usage{
select_samples(exp, ...)
}
\arguments{
\item{exp}{The \link{Experiment} object.}

\item{...}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> syntax for selecting columns.}
}
\value{
A new \link{Experiment} object.
}
\description{
Select columns of the sample information tibble of an \link{Experiment} object
using <\code{\link[=dplyr_tidy_select]{tidy-select}}> syntax.
For example, \code{select_samples(exp, group)} will return a new \link{Experiment} object
with a \code{sample_info} tibble having only the \code{sample} and \code{group} columns.
(Note that the \code{sample} column is always included automatically.)
}
\details{
Instead of \code{Experiment$select_samples()}, which updates the \link{Experiment} object in place,
this function returns a new \code{Experiment} object with \code{sample_info} tibble updated.
The original \code{Experiment} object will not be altered.
Besides, as the first argument is the \code{Experiment} object,
this function could be used in a \code{magrittr} pipe.

You should not, and don't need to, select or de-select the \code{sample} column.
The \code{sample} column is always included automatically.
}
