% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_variables}
\alias{select_variables}
\title{Select columns of the variable information tibble}
\usage{
select_variables(exp, ...)
}
\arguments{
\item{exp}{The \link{Experiment} object.}

\item{...}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> syntax for selecting columns.}
}
\value{
A new \link{Experiment} object.
}
\description{
Select columns of the variable information tibble of an \link{Experiment} object
using <\code{\link[=dplyr_tidy_select]{tidy-select}}> syntax.
For example, \code{select_variables(exp, type)} will return a new \link{Experiment} object
with a \code{var_info} tibble having only the \code{variable} and \code{type} columns.
(Note that the \code{variable} column is always included automatically.)
}
\details{
Instead of \code{Experiment$select_variables()}, which updates the \link{Experiment} object in place,
this function returns a new \code{Experiment} object with \code{var_info} tibble updated.
The original \code{Experiment} object will not be altered.
Besides, as the first argument is the \code{Experiment} object,
this function could be used in a \code{magrittr} pipe.

You should not, and don't need to, select or de-select the \code{variable} column.
The \code{variable} column is always included automatically.
}
