% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-data.R
\name{set_meta_data}
\alias{set_meta_data}
\alias{set_exp_type}
\alias{set_glycan_type}
\title{Set the meta data of an experiment}
\usage{
set_meta_data(exp, x, value)

set_exp_type(exp, value)

set_glycan_type(exp, value)
}
\arguments{
\item{exp}{An \code{\link[=experiment]{experiment()}}.}

\item{x}{A string,
the name of the meta data field.}

\item{value}{The value to set for the meta data field.}
}
\value{
The modified \code{\link[=experiment]{experiment()}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These metadata setters were deprecated with \code{experiment()}. Modify
\code{S4Vectors::metadata()} on a \code{SummarizedExperiment} instead.

Set meta data values for the experiment,
like the experiment type ("glycomics" or "glycoproteomics"),
or the glycan type ("N", "O", "O-GalNAc", "O-Man", "O-Fuc",
"O-GlcNAc", "O-Glc", "HMO", "GSL", "GAG", or "GPI").
}
\keyword{internal}
