% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyco-se.R
\name{show,GlycomicSE-method}
\alias{show,GlycomicSE-method}
\title{Show a GlycomicSE object}
\usage{
\S4method{show}{GlycomicSE}(object)
}
\arguments{
\item{object}{A \code{GlycomicSE} object.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Show a GlycomicSE object
}
