% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyco-se.R
\name{show,GlycoproteomicSE-method}
\alias{show,GlycoproteomicSE-method}
\title{Show a GlycoproteomicSE object}
\usage{
\S4method{show}{GlycoproteomicSE}(object)
}
\arguments{
\item{object}{A \code{GlycoproteomicSE} object.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Show a GlycoproteomicSE object
}
