% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split.glyexp_experiment}
\alias{split.glyexp_experiment}
\title{Split an Experiment}
\usage{
\method{split}{glyexp_experiment}(x, f, drop = FALSE, where = "var_info", ...)
}
\arguments{
\item{x}{An \code{\link[=experiment]{experiment()}}.}

\item{f}{<\code{\link[rlang:args_data_masking]{data-masking}}>
A column in \code{var_info} or \code{sample_info} that \code{as.factor(f)}defines the grouping.}

\item{drop}{Logical indicating if levels that do not occur should be dropped. Defaults to FALSE.}

\item{where}{Where to find the column, "var_info" or "sample_info".}

\item{...}{Ignored}
}
\value{
A named list of \code{\link[=experiment]{experiment()}} objects.
}
\description{
Devides an \code{\link[=experiment]{experiment()}} into a list of \code{\link[=experiment]{experiment()}} objects by \code{f}.
}
\examples{
split(toy_experiment, group, where = "sample_info")

}
