% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_experiment}
\alias{summarize_experiment}
\title{Identification overview}
\usage{
summarize_experiment(x, count_struct = NULL)
}
\arguments{
\item{x}{An \code{\link[=experiment]{experiment()}} object.}

\item{count_struct}{For counting glycopeptides and glycoforms.
whether to count the number of glycan structures or glycopeptides.
If \code{TRUE}, glycopeptides or glycoforms bearing different glycan structures
with the same glycan composition are counted as different ones.
If not provided (NULL), defaults to \code{TRUE} if \code{glycan_structure} column exists
in the variable information tibble, otherwise \code{FALSE}.}
}
\value{
A tibble with columns \code{item} and \code{n}
summarizing the results. The items include:
\itemize{
\item \code{composition}: The number of glycan compositions.
\item \code{structure}: The number of glycan structures.
\item \code{peptide}: The number of peptides.
\item \code{glycopeptide}: The number of unique combinations of peptides, sites, and glycans.
\item \code{glycoform}: The number of unique combinations of proteins, sites, and glycans.
\item \code{protein}: The number of proteins.
\item \code{glycosite}: The number of unique combinations of proteins and sites.
}
}
\description{
This function summarizes the number of
glycan compositions, glycan structures, glycopeptides, peptides,
glycoforms, glycoproteins, and glycosites in an \code{\link[=experiment]{experiment()}}.
}
\details{
The following columns are required in the variable information tibble:
\itemize{
\item \code{composition}: \code{glycan_composition}
\item \code{structure}: \code{glycan_structure}
\item \code{peptide}: \code{peptide}
\item \code{glycopeptide}: \code{glycan_composition} or \code{glycan_structure}, \code{peptide}, \code{peptide_site}
\item \code{glycoform}: \code{glycan_composition} or \code{glycan_structure},
\code{protein} or \code{proteins}, \code{protein_site} or \code{protein_sites}
\item \code{protein}: \code{protein} or \code{proteins}
\item \code{glycosite}: \code{protein} or \code{proteins}, \code{protein_site} or \code{protein_sites}
}

You can use \code{count_struct} parameter to control how to count glycopeptides and glycoforms,
either by glycan structures or by glycan compositions.
}
\examples{
exp <- real_experiment
summarize_experiment(exp)

}
