% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-arrange.R
\name{arrange_obs}
\alias{arrange_obs}
\alias{arrange_var}
\title{Arrange Sample or Variable Information}
\usage{
arrange_obs(exp, ...)

arrange_var(exp, ...)
}
\arguments{
\item{exp}{An \code{\link[=experiment]{experiment()}}.}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Variables to arrange by,
passed to \code{dplyr::arrange()} internally.}
}
\value{
An new \code{\link[=experiment]{experiment()}} object.
}
\description{
Arrange the sample or variable information tibble of an \code{\link[=experiment]{experiment()}}.

The same syntax as \code{dplyr::arrange()} is used.
For example, to arrange samples by the "group" column,
use \code{arrange_obs(exp, group)}.
This actually calls \code{dplyr::arrange()} on the sample information tibble
with the \code{group} column,
and then updates the expression matrix accordingly to match the new order.
}
\examples{
# Create a toy experiment for demonstration
exp <- toy_experiment()
# Add a type column to the variable information for demonstration
exp$var_info$type <- c("Y", "X", "Z", "Y")

# Arrange samples by group column
arranged_exp <- arrange_obs(exp, group)
arranged_exp$sample_info
arranged_exp$expr_mat

# Arrange variables by type column  
arranged_exp <- arrange_var(exp, type)
arranged_exp$var_info
arranged_exp$expr_mat

# Arrange by multiple columns
arrange_obs(exp, group, sample)$sample_info

}
