% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{create_experiment}
\alias{create_experiment}
\title{Function that Creates a New \link{Experiment} object}
\usage{
create_experiment(
  name = NA,
  expr_mat = NA,
  sample_info = NA,
  var_info = NA,
  rownames = FALSE
)
}
\arguments{
\item{name}{Name of the experiment.}

\item{expr_mat}{Expression matrix. Rows are variables and columns are samples.}

\item{sample_info}{Sample information, a tibble or a data.frame.
The first column should be "sample", corresponding to the column names of \code{expr_mat}.
The order of samples doesn't matter.}

\item{var_info}{Variable information, a tibble or a data.frame.
The first column should be "variable", corresponding to the row names of \code{expr_mat}.
The order of variables doesn't matter.}

\item{rownames}{Whether rownames of \code{sample_info} and \code{var_info} should be used.
If TRUE, the row names of \code{sample_info} and \code{sample_info} will be
converted to the first column (with names of "sample" and "variable", respectively).}
}
\value{
An \link{Experiment} object.
}
\description{
This function is a shortcut to \code{\link[=Experiment$new]{Experiment$new()}}.
}
\examples{
create_experiment("my_experiment", expr_mat, sample_info, var_info)
}
