% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.R
\name{dim.glyexp_experiment}
\alias{dim.glyexp_experiment}
\alias{dim<-.glyexp_experiment}
\title{Dimensions of an experiment}
\usage{
\method{dim}{glyexp_experiment}(x)

\method{dim}{glyexp_experiment}(x) <- value
}
\arguments{
\item{x}{An experiment object.}

\item{value}{Ignored.}
}
\value{
A vector with two elements: the number of variables and the number of samples.
}
\description{
Retrieve the dimensions of an experiment object,
i.e. the number of variables and samples.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This legacy API is retained temporarily for compatibility. Use
\code{\link[=GlycomicSE]{GlycomicSE()}} or \code{\link[=GlycoproteomicSE]{GlycoproteomicSE()}} as the default data container.
}
\keyword{internal}
