% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{Experiment}
\alias{Experiment}
\title{R6 Class for a Glycoproteomics or Glycomics Experiment}
\description{
A class managing expression matrix, sample information, and variable information.
}
\details{
This class synchronizes three data types commonly used in a glycproteomics
or glycomics research: the expression matrix, the sample information, and
the variable information. It helps ensure that the three datasets altered
in sync. For example, filtering samples based on some conditions will
automatically affect both the expression matrix and the sample information.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Name of the experiment.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Experiment-new}{\code{Experiment$new()}}
\item \href{#method-Experiment-get_expr_mat}{\code{Experiment$get_expr_mat()}}
\item \href{#method-Experiment-get_sample_info}{\code{Experiment$get_sample_info()}}
\item \href{#method-Experiment-get_var_info}{\code{Experiment$get_var_info()}}
\item \href{#method-Experiment-filter_samples}{\code{Experiment$filter_samples()}}
\item \href{#method-Experiment-filter_variables}{\code{Experiment$filter_variables()}}
\item \href{#method-Experiment-mutate_samples}{\code{Experiment$mutate_samples()}}
\item \href{#method-Experiment-mutate_variables}{\code{Experiment$mutate_variables()}}
\item \href{#method-Experiment-select_samples}{\code{Experiment$select_samples()}}
\item \href{#method-Experiment-select_variables}{\code{Experiment$select_variables()}}
\item \href{#method-Experiment-clone}{\code{Experiment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Experiment-new"></a>}}
\if{latex}{\out{\hypertarget{method-Experiment-new}{}}}
\subsection{Method \code{new()}}{
Create a new experiment object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$new(
  name = NA,
  expr_mat = NA,
  sample_info = NA,
  var_info = NA,
  rownames = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the experiment.}

\item{\code{expr_mat}}{Expression matrix. Rows are variables and columns are samples.}

\item{\code{sample_info}}{Sample information, a tibble or a data.frame.
The first column should be "sample", corresponding to the column names of \code{expr_mat}.
The order of samples doesn't matter.}

\item{\code{var_info}}{Variable information, a tibble or a data.frame.
The first column should be "variable", corresponding to the row names of \code{expr_mat}.
The order of variables doesn't matter.}

\item{\code{rownames}}{Whether rownames of \code{sample_info} and \code{var_info} should be used.
If TRUE, the row names of \code{sample_info} and \code{sample_info} will be
converted to the first column (with names of "sample" and "variable", respectively).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Experiment-get_expr_mat"></a>}}
\if{latex}{\out{\hypertarget{method-Experiment-get_expr_mat}{}}}
\subsection{Method \code{get_expr_mat()}}{
Get a copy of the expression matrix.
The rows are variables and the columns are samples.
A function \code{\link[=get_expr_mat]{get_expr_mat()}} is also provided for this purpose.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_expr_mat()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A copy of the expression matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Experiment-get_sample_info"></a>}}
\if{latex}{\out{\hypertarget{method-Experiment-get_sample_info}{}}}
\subsection{Method \code{get_sample_info()}}{
Get a copy of the sample information tibble.
This is a tibble with the first conserved column "sample",
and the rest columns as meta data of samples,
e.g. "group", "batch", etc.
The "sample" column equals to the column names of \code{expr_mat}.
A function \code{\link[=get_sample_info]{get_sample_info()}} is also provided for this purpose.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_sample_info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A copy of the sample information tibble.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Experiment-get_var_info"></a>}}
\if{latex}{\out{\hypertarget{method-Experiment-get_var_info}{}}}
\subsection{Method \code{get_var_info()}}{
Get a copy of the variable information tibble.
This is a tibble with the first conserved column "variable",
and the rest columns as meta data of variables,
e.g. "peptide", "protein", "glycan_composition", etc.
The "variable" column equals to the row names of \code{expr_mat}.
A function \code{\link[=get_var_info]{get_var_info()}} is also provided for this purpose.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_var_info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A copy of the variable information tibble.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Experiment-filter_samples"></a>}}
\if{latex}{\out{\hypertarget{method-Experiment-filter_samples}{}}}
\subsection{Method \code{filter_samples()}}{
This method filters samples based on conditions specified in the \code{\link[dplyr:filter]{dplyr::filter()}} function.
For example, \code{filter_samples(group == "A")} will keep samples with "group" equal to "A".
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$filter_samples(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Conditions for filtering samples, passed to \code{\link[dplyr:filter]{dplyr::filter()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This will affect both the expression matrix and the sample information tibble.
The \link{Experiment} object will be updated in place.
The \link{Experiment} object will be returned invisibly to allow chaining.
}

\subsection{Returns}{
The \link{Experiment} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Experiment-filter_variables"></a>}}
\if{latex}{\out{\hypertarget{method-Experiment-filter_variables}{}}}
\subsection{Method \code{filter_variables()}}{
This method filters variables based on conditions specified in the \code{\link[dplyr:filter]{dplyr::filter()}} function.
For example, \code{filter_variables(type == "B")} will keep variables with "type" equal to "B".
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$filter_variables(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Conditions for filtering variables, passed to \code{\link[dplyr:filter]{dplyr::filter()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This will affect both the expression matrix and the variable information tibble.
The \link{Experiment} object will be updated in place.
The \link{Experiment} object will be returned invisibly to allow chaining.
}

\subsection{Returns}{
The \link{Experiment} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Experiment-mutate_samples"></a>}}
\if{latex}{\out{\hypertarget{method-Experiment-mutate_samples}{}}}
\subsection{Method \code{mutate_samples()}}{
This function mutates the sample information tibble using \code{\link[dplyr:mutate]{dplyr::mutate()}}.
For example, \verb{mutate_samples(new_group = if_else(new_col = 1)}
will add a new column "new_group" to the sample information tibble.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$mutate_samples(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Mutations for the sample information tibble, passed to \code{\link[dplyr:mutate]{dplyr::mutate()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The \link{Experiment} object will be updated in place.
The \link{Experiment} object will be returned invisibly to allow chaining.

The columns "sample" is protected from being modified or renamed.
It is essential for keeping the linkage between the sample information and the expression matrix.
You may add new columns by transforming the "sample" column.
}

\subsection{Returns}{
The \link{Experiment} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Experiment-mutate_variables"></a>}}
\if{latex}{\out{\hypertarget{method-Experiment-mutate_variables}{}}}
\subsection{Method \code{mutate_variables()}}{
This function mutates the variable information tibble using \code{\link[dplyr:mutate]{dplyr::mutate()}}.
For example, \verb{mutate_variables(new_type = if_else(new_col = 1)}
will add a new column "new_type" to the variable information tibble.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$mutate_variables(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Mutations for the variable information tibble, passed to \code{\link[dplyr:mutate]{dplyr::mutate()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The \link{Experiment} object will be updated in place.
The \link{Experiment} object will be returned invisibly to allow chaining.

The columns "variable" is protected from being modified or renamed.
It is essential for keeping the linkage between the variable information and the expression matrix.
You may add new columns by transforming the "variable" column.
}

\subsection{Returns}{
The \link{Experiment} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Experiment-select_samples"></a>}}
\if{latex}{\out{\hypertarget{method-Experiment-select_samples}{}}}
\subsection{Method \code{select_samples()}}{
This function selects columns in the sample information tibble.
You don't need to select the "sample" column explicitly, as it is always selected.
For example, \code{exp$select_samples(group)} will result in a tibble
with the "sample" and "group" columns.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$select_samples(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> for selecting columns, passed to \code{\link[dplyr:select]{dplyr::select()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The \link{Experiment} object will be updated in place.
The \link{Experiment} object will be returned invisibly to allow chaining.

When the user tries to select or deselect the "sample" column explicitly,
an error will be thrown.
}

\subsection{Returns}{
The \link{Experiment} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Experiment-select_variables"></a>}}
\if{latex}{\out{\hypertarget{method-Experiment-select_variables}{}}}
\subsection{Method \code{select_variables()}}{
This function selects columns in the variable information tibble.
You don't need to select the "variable" column explicitly, as it is always selected.
For example, \code{exp$select_variables(peptide, protein)} will result in a tibble
with the "variable", "peptide", and "protein" columns.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$select_variables(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> for selecting columns, passed to \code{\link[dplyr:select]{dplyr::select()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The \link{Experiment} object will be updated in place.
The \link{Experiment} object will be returned invisibly to allow chaining.

When the user tries to select or deselect the "variable" column explicitly,
an error will be thrown.
}

\subsection{Returns}{
The \link{Experiment} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Experiment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Experiment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
