% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{get_expr_mat}
\alias{get_expr_mat}
\title{Get the expression matrix}
\usage{
get_expr_mat(exp)
}
\arguments{
\item{exp}{The \link{Experiment} object.}
}
\value{
The expression matrix.
}
\description{
Get the expression matrix from an \link{Experiment} object.
}
\details{
A copy, instead of the original matrix, will be returned.
This prevents the original matrix from being altered.
Alteration of data in an \link{Experiment} object is strictly managed
by the class, so that the expression matrix, sample information,
and variable information are always consistent.
}
