% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-getters.R
\name{get_expr_mat}
\alias{get_expr_mat}
\title{Get the expression matrix of an experiment}
\usage{
get_expr_mat(exp)
}
\arguments{
\item{exp}{An \code{\link[=experiment]{experiment()}}.}
}
\value{
A matrix of expression values.
}
\description{
A \code{matrix} of expression values with samples as columns and variables as rows.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This legacy API is retained temporarily for compatibility. Use
\code{\link[=GlycomicSE]{GlycomicSE()}} or \code{\link[=GlycoproteomicSE]{GlycoproteomicSE()}} as the default data container.
}
\keyword{internal}
