% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-getters.R
\name{get_expr_mat}
\alias{get_expr_mat}
\title{Get the expression matrix of an experiment}
\usage{
get_expr_mat(exp)
}
\arguments{
\item{exp}{An \code{\link[=experiment]{experiment()}}.}
}
\value{
A matrix of expression values.
}
\description{
A \code{matrix} of expression values with samples as columns and variables as rows.
}
