% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-getters.R
\name{get_meta_data}
\alias{get_meta_data}
\title{Get the meta data of an \code{\link[=experiment]{experiment()}}}
\usage{
get_meta_data(exp)
}
\arguments{
\item{exp}{An \code{\link[=experiment]{experiment()}}.}
}
\value{
The meta data as a named list.
}
\description{
Meta data is a named list of arbitrary data about the experiment,
e.g. "structure_type", "glycan_type", and "data_type".
}
