% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-getters.R
\name{get_meta_data}
\alias{get_meta_data}
\title{Get the meta data of an \code{\link[=experiment]{experiment()}}}
\usage{
get_meta_data(exp, x)
}
\arguments{
\item{exp}{An \code{\link[=experiment]{experiment()}}.}

\item{x}{A string, the name of the meta data field.}
}
\value{
The value of the meta data field. If the field does not exist,
\code{NULL} will be returned.
}
\description{
Meta data is a named list of arbitrary data about the experiment,
e.g. "exp_type", "glycan_type", etc.
}
