% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-data.R
\name{get_meta_data}
\alias{get_meta_data}
\alias{get_exp_type}
\alias{get_glycan_type}
\title{Get the meta data of an \code{\link[=experiment]{experiment()}}}
\usage{
get_meta_data(exp, x)

get_exp_type(exp)

get_glycan_type(exp)
}
\arguments{
\item{exp}{An \code{\link[=experiment]{experiment()}}.}

\item{x}{A string, the name of the meta data field.}
}
\value{
The value of the meta data field. If the field does not exist,
\code{NULL} will be returned.
}
\description{
Meta data is some descriptions about the experiment,
like the experiment type ("glycomics" or "glycoproteomics"),
or the glycan type ("N" or "O").
}
