% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-getters.R
\name{get_sample_info}
\alias{get_sample_info}
\title{Get the sample information of an experiment}
\usage{
get_sample_info(exp)
}
\arguments{
\item{exp}{An \code{\link[=experiment]{experiment()}}.}
}
\value{
A tibble of sample information.
}
\description{
A \code{tibble} of sample information, with the first column being "sample".
}
