% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-getters.R
\name{get_var_info}
\alias{get_var_info}
\title{Get the Variable Information of an \code{\link[=experiment]{experiment()}}}
\usage{
get_var_info(exp)
}
\arguments{
\item{exp}{An \code{\link[=experiment]{experiment()}}.}
}
\value{
A tibble of variable information.
}
\description{
A \code{tibble} of variable information, with the first column being "variable".
}
