% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyco-se.R
\name{GlycomicSE}
\alias{GlycomicSE}
\title{Create a GlycomicSE object}
\usage{
GlycomicSE(abundance, ...)
}
\arguments{
\item{abundance}{A numeric abundance matrix with glycans as rows and samples
as columns.}

\item{...}{Arguments passed to \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}},
such as \code{rowData}, \code{colData}, and \code{metadata}.}
}
\value{
A \code{GlycomicSE} object.
}
\description{
Create a GlycomicSE object
}
