% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyco-se.R
\name{GlycoproteomicSE}
\alias{GlycoproteomicSE}
\title{Create a GlycoproteomicSE object}
\usage{
GlycoproteomicSE(abundance, ...)
}
\arguments{
\item{abundance}{A numeric abundance matrix with glycopeptides or glycoforms
as rows and samples as columns.}

\item{...}{Arguments passed to \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}},
such as \code{rowData}, \code{colData}, and \code{metadata}.}
}
\value{
A \code{GlycoproteomicSE} object.
}
\description{
Create a GlycoproteomicSE object
}
