% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyexp-package.R
\docType{package}
\name{glyexp-package}
\alias{glyexp}
\alias{glyexp-package}
\title{glyexp: Glycoproteomics and Glycomics Experiments}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a tidy data framework for managing glycoproteomics and glycomics experimental data. The 'GlycomicSE' and 'GlycoproteomicSE' classes extend 'SummarizedExperiment' with validated glycomics and glycoproteomics schemas. The package provides dplyr-style data manipulation functions (filter, mutate, select, arrange, slice, join) for seamless data wrangling and retains the legacy 'experiment()' container temporarily for compatibility. As the data core of the 'glycoverse' ecosystem, it provides consistent interfaces for data exchange and analysis workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://glycoverse.github.io/glyexp/}
  \item \url{https://github.com/glycoverse/glyexp}
  \item Report bugs at \url{https://github.com/glycoverse/glyexp/issues}
}

}
\author{
\strong{Maintainer}: Bin Fu \email{23110220018@m.fudan.edu.cn} (\href{https://orcid.org/0000-0001-8567-2997}{ORCID}) [copyright holder]

}
\keyword{internal}
