% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyexp-package.R
\docType{package}
\name{glyexp-package}
\alias{glyexp}
\alias{glyexp-package}
\title{glyexp: Glycoproteomics and Glycomics Experiments}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides classes and functions to manage data in glycoproteomics and glycomics experiments, including expression matrix, variable information (peptide, protein, glycan composition, etc.), and sample information (group, batch, etc.).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://glycoverse.github.io/glyexp/}
}

}
\author{
\strong{Maintainer}: Bin Fu \email{23110220018@m.fudan.edu.cn} (\href{https://orcid.org/0000-0001-8567-2997}{ORCID})

}
\keyword{internal}
