% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-and-variables.R
\name{n_samples}
\alias{n_samples}
\alias{n_variables}
\title{Get Number of Samples or Variables of an Experiment}
\usage{
n_samples(exp)

n_variables(exp)
}
\arguments{
\item{exp}{An \code{\link[=experiment]{experiment()}}.}
}
\value{
An integer with the number of samples or variables.
}
\description{
Getting the number of samples or variables of an \code{\link[=experiment]{experiment()}}.
Syntax sugar for \code{ncol(exp$expr_mat)} and \code{nrow(exp$expr_mat)}.
}
\examples{
exp <- toy_experiment()
n_samples(exp)
n_variables(exp)

}
