% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-select.R
\name{select_obs}
\alias{select_obs}
\alias{select_var}
\title{Select Columns of the Sample or Variable Information Tibble}
\usage{
select_obs(exp, ...)

select_var(exp, ...)
}
\arguments{
\item{exp}{An \code{\link[=experiment]{experiment()}}.}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Column names to select.
If empty, all columns except the \code{sample} or \code{variable} column will be discarded.}
}
\value{
An new \code{\link[=experiment]{experiment()}} object.
}
\description{
These two functions provide a way to trimming down the sample or variable information tibble
of an \code{\link[=experiment]{experiment()}} to only the columns of interest.

The same syntax as \code{dplyr::select()} is used.
For example, to get a new \code{\link[=experiment]{experiment()}} with only the "sample" and "group"
columns in the sample information tibble,
use \code{select_obs(exp, group)}.
Note that you don't need to (and you can't) explicitly select or deselect the
\code{sample} column in \code{sample_info}.
It is automatically handled by \code{select_obs()}, always being selected.
The same applies to the \code{variable} column in \code{var_info}.
}
\examples{
library(magrittr)

toy_exp <- toy_experiment()

toy_exp_2 <- toy_exp \%>\%
  select_obs(group) \%>\%
  select_var(protein, peptide)

get_sample_info(toy_exp_2)
get_var_info(toy_exp_2)

}
