% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-data.R
\name{set_meta_data}
\alias{set_meta_data}
\alias{set_exp_type}
\alias{set_glycan_type}
\title{Set the meta data of an experiment}
\usage{
set_meta_data(exp, x, value)

set_exp_type(exp, value)

set_glycan_type(exp, value)
}
\arguments{
\item{exp}{An \code{\link[=experiment]{experiment()}}.}

\item{x}{A string,
the name of the meta data field.}

\item{value}{The value to set for the meta data field.}
}
\value{
The modified \code{\link[=experiment]{experiment()}} object.
}
\description{
Set meta data values for the experiment,
like the experiment type ("glycomics" or "glycoproteomics"),
or the glycan type ("N", "O-GalNAc", "O-GlcNAc", "O-Man", "O-Fuc", or "O-Glc").
}
