% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{[.glyexp_experiment}
\alias{[.glyexp_experiment}
\alias{[<-.glyexp_experiment}
\title{Subsetting experiments}
\usage{
\method{[}{glyexp_experiment}(x, i, j, name = NULL, ...)

\method{[}{glyexp_experiment}(x, i, j, ...) <- value
}
\arguments{
\item{x}{An \code{\link[=experiment]{experiment()}}.}

\item{i, j}{Row (variable) and column (sample) indices to subset.}

\item{name}{Name of the subsetted experiment.}

\item{...}{Ignored.}

\item{value}{Ignored.}
}
\value{
An \code{\link[=experiment]{experiment()}} object.
}
\description{
Getting a subset of an experiment object.
Subsetting is first done on the expression matrix,
then the sample information and variable information tibbles are filtered
and ordered accordingly.

Syntax for \code{[} is similar to subsetting a matrix,
with some differences:
\itemize{
\item Both row and column indices are required, i.e. \code{exp[i]} is not allowed,
but \code{exp[i, ]} and \code{exp[, j]} are allowed.
\item \code{drop} argument is not supported.
Subsetting an experiment always returns an new experiment,
even if it has only one sample or one variable.
\item A \code{name} argument is added to change the name of the subsetted experiment.
If omitted, the name of the original experiment will be used.
}

Assigning to a subset of an experiment is not allowed,
i.e., \verb{exp[1, 1[ <- 0} will raise an error.
You can create a new experiment with new data if needed.
}
\examples{
# Create a toy experiment for demonstration
expr_mat <- matrix(1:9, nrow = 3)
colnames(expr_mat) <- c("S1", "S2", "S3")
rownames(expr_mat) <- c("V1", "V2", "V3")
sample_info <- tibble::tibble(
  sample = c("S1", "S2", "S3"),
  group = rep("A", 3)
)
var_info <- tibble::tibble(
  variable = c("V1", "V2", "V3"),
  type = rep("B", 3)
)
exp <- experiment("my_exp", expr_mat, sample_info, var_info)

# Subsetting single samples
exp[, "S1"]
exp[, 1]

# Subsetting single variables
exp["V1", ]
exp[1, ]

# Subsetting multiple samples and variables
exp[c("V1", "V2"), c("S2", "S3")]
exp[c(1, 2), c(2, 3)]

# Subsetting with a new name
exp[c("V1", "V2"), c("S2", "S3"), name = "sub_exp"]

# Create a copy
exp[, ]

}
