% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_experiment}
\alias{toy_experiment}
\title{Toy experiment}
\format{
\subsection{\code{toy_experiment}}{

An \code{\link[=experiment]{experiment()}} object with 4 variables and 6 samples.
\code{var_info} contains fields: \code{protein}, \code{peptide}, and \code{glycan_composition}.
\code{sample_info} contains fields: \code{group} and \code{batch}.
}
}
\usage{
toy_experiment
}
\description{
A toy experiment object for new users to play with
to get familiar with \code{\link[=experiment]{experiment()}} objects.
}
\keyword{datasets}
